kernel: Use defconfig instead of full fledged kernel configuration
Rely on the Kconfig defconfig mechanism to fill all the missing options,
instead of needing to set them all in the kernel configurations like what was
previously done.

This will allow to trim down a lot the configuration files, avoid carrying
unused configuration options and preserve the developers mental health.

Signed-off-by: Maxime Ripard <maxime.ripard@free-electrons.com>