mvebu: Add sub-profiles
In order to support the various board NAND layout that we support, introduce a
sub-profile similar to the one used by ar71xx.

These subprofiles provide a default implementation for most of the building
functions, while allowing each sub-profile to override any of these operations
in order to have a more specific behaviour, like Mamba expects for example.

Signed-off-by: Maxime Ripard <maxime.ripard@free-electrons.com>